+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for opening DICOMDIR files using `open_dicomdir()`.
+
 ## [0.13.0] - 2023-11-11
 
 ### Added